Add dependencies to layouthelper and ios descriptor pods (#2140)
Summary:
This diff makes sure that pod lib lint succeeds and the CI is green for iOS.

It got broken when I landed my changes for the refactoring of FlipperKitLayoutPlugin. It wasn't picked up by flipper-oss-ios-job as it builds sample app and sample app had all the paths to the header search path which individually FlipperKitLayoutHelpers and FlipperKitIOSDescriptors were missing.

I will add the validation of pods as a job in sandcastle when one touches FlipperKit pods, this will avoid such cases in future.

## Changelog

Pull Request resolved: https://github.com/facebook/flipper/pull/2140

Test Plan:
Imported from GitHub, without a `Test Plan:` line.

iOS jobs are green on CI in GH.

Reviewed By: passy

Differential Revision: D27417743

Pulled By: priteshrnandgaonkar

fbshipit-source-id: 17aa4b04bf9954c4bd138261103751a245409568