Update yarn.lock for ReactNativeFlipperExample
Summary:
Recently our release script started to bump "react-native-flipper" dependency. The problem is that yarn.lock is not updated when the version bumped, so they lost sync with package.json and build on github are failing because of that: https://github.com/facebook/flipper/pull/2172/checks?check_run_id=2298828522.

In this diff I'm just updating yarn.lock to unblock builds, need to also address bumping script to avoid this issue in next release.

Reviewed By: mweststrate

Differential Revision: D27654577

fbshipit-source-id: 4ec0f6899460a92f06e215cf70df39e4bdbbfe03