Fix re-loading of source maps with "-fast-refresh" flag enabled
Summary: Fixed loading of source maps when dev server is running with --fast-refresh flag set. Looks like it was broken after last metro upgrade because it changed the way how module paths are passed to dev server. We also had a hack to workaround bug when requested module paths were malformed, but it looks this hack is not needed now, so I deleted it as well.

Reviewed By: passy

Differential Revision: D28120080

fbshipit-source-id: 03a39c9840da8f4c8c72be6189c0b71c62fac7f0