Re-integrate bug report traces after Sandy Migration
Summary:
The migration to Sandy (D28060888) breaks the existing integration we have with importing Flipper traces from bug reports. Given that, I broke this diff out from that one just to highlight the changes.

`pluginStates2` appears to be the field to use for Sandy traces (rather than `store`, which we were using before)

Differential Revision: D28209561

fbshipit-source-id: 9803e1a60b1b1deba6403b50b4079aac2a19efa6