Bump all Pods to latest (#2275)
Summary:
Pull Request resolved: https://github.com/facebook/flipper/pull/2275

Forced all current pods to use the latest available. If this builds successfully, will create a PR against React Native so that they become the default, and remove them here again.

This job currently does limit the architectures we build for. Pritesh is investigating if we can wide that again, as this might not be acceptable to RN in general.

This doesn't bump the android version to 0.88, as the artifact publishing failed for that version, but since there are currently less problems with the Android version I wanted to go ahead with just iOS.

Reviewed By: cekkaewnumchai, priteshrnandgaonkar

Differential Revision: D28257283

fbshipit-source-id: a631b1c354617ed356cd1e74b9e8e8aa050eed60