"Fix" "typo" that's blocking Android releases
Summary:
Bit of an odd one. Our publishing plugin changed what fields
it expect to define the license of the project. That has caused
Maven Central to reject our uploads:

https://github.com/vanniktech/gradle-maven-publish-plugin/blob/af0de22a5a7ee2efcbaa77756769392e13dd6b38/src/main/kotlin/com/vanniktech/maven/publish/legacy/BaseSetup.kt#L70

I hope this is a temporary "fix".

Test Plan:
Uploaded to my local maven m2 and now the fields are back.