Attribute error reports with plugin id and version
Summary: This diff adds attributing of all error reports with the information about the plugins caused them. To do that we try to map error stack trace to plugins by searching for plugin directories in it. We can do this only for plugins loaded from disk as we are looking for their physical locations in stack. To make it possible to search for multiple plugins in one go we build and maintain a trie containing all the loaded plugins locations.

Reviewed By: passy

Differential Revision: D28475938

fbshipit-source-id: 59c8ac39ffe54262b1d57f21fad44eb89748dff2