Update docs workflow (#2342)
Summary:
Updated to the latest version which apparently doesn't need a privileged token anymore, which would be good news.
Also updated the message to include a `[ci skip]` which instructs CircleCI not to run because we're having a chicken-and-egg-problem here with CircleCI otherwise needing a config to tell it not to run, otherwise it will run and fail because there's no config.

Pull Request resolved: https://github.com/facebook/flipper/pull/2342

Test Plan:
hopeitwork

Task: T91157540

Reviewed By: mweststrate

Differential Revision: D28539468

Pulled By: passy

fbshipit-source-id: 609c3fa1578e0506aadeb225c4e38ff53db67934