docs: update network-plugin.mdx android (#2368)
Summary:
Flipper network plugin does not work without the above added lines.

Without the added lines, the network plugin simply does not work on Android with Flipper version 0.91.0. I found this solution in your example react-native project. This should be stated in the docs.

## Changelog

This should not be stated in changelog.

Pull Request resolved: https://github.com/facebook/flipper/pull/2368

Test Plan: There is no code.

Reviewed By: passy

Differential Revision: D28714451

Pulled By: nikoant

fbshipit-source-id: 81583fcf590e1c603d67bcb649067393455549cc