Remove minimist pin
Summary:
I'm going through our resolutions one-by-one, removing them and then seeing what the resulting change in the yarn.lock is and if it falls below what was pinned prior to D28835488.

In this case, we're above the previous lock which means we can safely remove it without getting a vulnerable version back.

Reviewed By: muraziz

Differential Revision: D28835545

fbshipit-source-id: 3159e1a165bbd4310c44520b0fe861a0d3a74fed