Add github spec repo as a source in RN Sample app (#2412)
Summary:
This diff adds the cocoapod's github repo as a source otherwise Cocoapod uses the cdn to refer the latest pods. We need this as in create_pr job we update the Podfile.lock of this RN sample project and it fails to find the latest flipper pod because Cocapod takes a while to update its cdn. Reffering the specs directly from github repo solves the issue.
## Changelog

Add github spec repo as a source in the Podfile.

Pull Request resolved: https://github.com/facebook/flipper/pull/2412

Test Plan:
CI
https://github.com/facebook/flipper/actions/runs/902547740

Reviewed By: nikoant

Differential Revision: D28896796

Pulled By: priteshrnandgaonkar

fbshipit-source-id: d5a2f10d42fa91218f853cdfb1920a10f438c697