Better error for ports in use
Summary:
If port 8088 / 8089 are in use, devices will not be able to connect to Flipper. We have checks for this in place, however it turns out that users can miss them. Potentially this can be caused for mulltiple reasons:

1. The notification is shown fairly short (in my testing it disappeared sometimes in < 2 secs
2. The notification is shown outside of the Flipper area
3. Flipper might not have permissions to show notifs in the first place.

Changed the notification to show inside Flipper, not hide automatically, and added clearer instructions. Including mentioning the port number (since there are multiple)

Changelog: [Flipper] Provide clearer error if Flipper ports are already in use, making device connections impossible

Differential Revision: D29128014

fbshipit-source-id: 16d9b451aa84281744155bfa1042429911a7b774