Enable configuring OnDemand for testing
Summary: While working on D29694696, noticed it is tricky to test against an OnDemand, as the URL format is pretty specific and authentication flow uses an unrelated URL base constant. Introduced `FB_ONDEMAND` env var to make testing easier. If this is a common scenario we could consider making it a setting in the long run.

Reviewed By: timur-valiev

Differential Revision: D29694707

fbshipit-source-id: 747ceb80c95c3bdf0276497c41c9ff5ed9a198f2